+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Use `del-cli` executable instead of `del` in order to not interfere with system commands on Windows ([#7](https://github.com/smartive/generator-giuseppe/issues/7))
+- Updated to new @types for type definitions.
 
 ## [0.2.0]
 #### Added